# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Duolingo Clone API",
    "description": "Duolingo clone backend API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/courses/{course_id}/path": {
      "get": {
        "tags": [
          "Courses"
        ],
        "summary": "Get Course Path",
        "description": "Retrieve full hierarchical learning tree (units -> skills -> lessons)\nwith user completion and locking.",
        "operationId": "get_course_path_api_courses__course_id__path_get",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Course Id"
            }
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoursePathResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/attempts/{attempt_id}/check": {
      "post": {
        "tags": [
          "Attempts"
        ],
        "summary": "Check Answer",
        "description": "Authoritatively check a submitted exercise answer against server answer keys.",
        "operationId": "check_answer_api_attempts__attempt_id__check_post",
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Attempt Id"
            }
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckAnswerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckAnswerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/attempts/{attempt_id}/complete": {
      "post": {
        "tags": [
          "Attempts"
        ],
        "summary": "Complete Attempt",
        "description": "Authoritatively complete an attempt, deriving accuracy and awarding XP.",
        "operationId": "complete_attempt_api_attempts__attempt_id__complete_post",
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Attempt Id"
            }
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompleteAttemptResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/attempts/{attempt_id}/abandon": {
      "post": {
        "tags": [
          "Attempts"
        ],
        "summary": "Abandon Attempt",
        "description": "Abandon an active lesson attempt.",
        "operationId": "abandon_attempt_api_attempts__attempt_id__abandon_post",
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Attempt Id"
            }
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AbandonAttemptResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/debug/advance-day": {
      "post": {
        "tags": [
          "Debug"
        ],
        "summary": "Advance Day",
        "description": "Advance the user's logical clock by N days to simulate time progression.",
        "operationId": "advance_day_api_debug_advance_day_post",
        "parameters": [
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvanceDayRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdvanceDayResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/debug/unlock-all": {
      "post": {
        "tags": [
          "Debug"
        ],
        "summary": "Unlock All",
        "description": "Debug helper: unlock all lessons and skills for user 1.",
        "operationId": "unlock_all_api_debug_unlock_all_post",
        "parameters": [
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleMessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/debug/reset-demo": {
      "post": {
        "tags": [
          "Debug"
        ],
        "summary": "Reset Demo",
        "description": "Debug helper: reset user 1 progress, hearts, gems, streak, and restore seed baseline.",
        "operationId": "reset_demo_api_debug_reset_demo_post",
        "parameters": [
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleMessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AbandonAttemptResponse": {
        "properties": {
          "attempt_id": {
            "type": "integer",
            "title": "Attempt Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "attempt_id",
          "status"
        ],
        "title": "AbandonAttemptResponse"
      },
...
      "AdvanceDayRequest": {
        "properties": {
          "days": {
            "type": "integer",
            "title": "Days",
            "default": 1
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AdvanceDayRequest"
      },
      "AdvanceDayResponse": {
        "properties": {
          "date_offset_days": {
            "type": "integer",
            "title": "Date Offset Days"
          },
          "logical_now": {
            "type": "string",
            "title": "Logical Now"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "date_offset_days",
          "logical_now"
        ],
        "title": "AdvanceDayResponse"
      },
      "CheckAnswerRequest": {
        "properties": {
          "exercise_id": {
            "type": "integer",
            "title": "Exercise Id"
          },
          "user_answer": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "User Answer"
          },
          "answer": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Answer"
          }
        },
        "type": "object",
        "required": [
          "exercise_id"
        ],
        "title": "CheckAnswerRequest"
      },
      "CheckAnswerResponse": {
        "properties": {
          "is_correct": {
            "type": "boolean",
            "title": "Is Correct"
          },
          "correct": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correct"
          },
          "correct_answer": {
            "title": "Correct Answer"
          },
          "hearts_remaining": {
            "type": "integer",
            "title": "Hearts Remaining"
          },
          "hearts": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hearts"
          },
          "attempt_status": {
            "type": "string",
            "title": "Attempt Status"
          },
          "lesson_failed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lesson Failed",
            "default": false
          },
          "explanation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explanation"
          }
        },
        "type": "object",
        "required": [
          "is_correct",
          "correct_answer",
          "hearts_remaining",
          "attempt_status"
        ],
        "title": "CheckAnswerResponse"
      },
      "CompleteAttemptResponse": {
        "properties": {
          "attempt_id": {
            "type": "integer",
            "title": "Attempt Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "xp_earned": {
            "type": "integer",
            "title": "Xp Earned"
          },
          "accuracy": {
            "type": "number",
            "title": "Accuracy"
          },
          "hearts_lost": {
            "type": "integer",
            "title": "Hearts Lost"
          },
          "is_replay": {
            "type": "boolean",
            "title": "Is Replay"
          },
          "new_crown_earned": {
            "type": "boolean",
            "title": "New Crown Earned"
          },
          "streak_current": {
            "type": "integer",
            "title": "Streak Current"
          },
          "time_spent_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Spent Seconds"
          },
          "crowns_after": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crowns After"
          },
          "skill_completed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skill Completed"
          },
          "unit_completed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Completed"
          },
          "new_achievements": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "New Achievements"
          },
          "streak_extended": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Streak Extended"
          }
        },
        "type": "object",
        "required": [
          "attempt_id",
          "status",
          "xp_earned",
          "accuracy",
          "hearts_lost",
          "is_replay",
          "new_crown_earned",
          "streak_current"
        ],
        "title": "CompleteAttemptResponse"
      },
...
      "CoursePathResponse": {
        "properties": {
          "course_id": {
            "type": "integer",
            "title": "Course Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "units": {
            "items": {
              "$ref": "#/components/schemas/UnitSummary"
            },
            "type": "array",
            "title": "Units"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "course_id",
          "title",
          "units"
        ],
        "title": "CoursePathResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LessonSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "lesson_order": {
            "type": "integer",
            "title": "Lesson Order"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "is_completed": {
            "type": "boolean",
            "title": "Is Completed"
          },
          "is_locked": {
            "type": "boolean",
            "title": "Is Locked"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "lesson_order",
          "title",
          "is_completed",
          "is_locked"
        ],
        "title": "LessonSummary"
      },
...
      "SimpleMessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message"
        ],
        "title": "SimpleMessageResponse"
      },
      "SkillSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "skill_order": {
            "type": "integer",
            "title": "Skill Order"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "icon_name": {
            "type": "string",
            "title": "Icon Name"
          },
          "total_crowns": {
            "type": "integer",
            "title": "Total Crowns"
          },
          "crowns_earned": {
            "type": "integer",
            "title": "Crowns Earned"
          },
          "lessons": {
            "items": {
              "$ref": "#/components/schemas/LessonSummary"
            },
            "type": "array",
            "title": "Lessons"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "skill_order",
          "name",
          "description",
          "icon_name",
          "total_crowns",
          "crowns_earned",
          "lessons"
        ],
        "title": "SkillSummary"
      },
...
      "UnitSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "unit_order": {
            "type": "integer",
            "title": "Unit Order"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/SkillSummary"
            },
            "type": "array",
            "title": "Skills"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "unit_order",
          "title",
          "description",
          "skills"
        ],
        "title": "UnitSummary"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}